fix(sse): apply client-abort guard to sibling HTTP servers (apiBridge, liveWS, embedWS)
Extends #fix-dev-server-aborted beyond the Next dev/prod server to the
other Node http servers that are equally exposed to a client aborting the
TCP socket (navigation/HMR/bfcache) while a response is in flight:

- src/lib/apiBridgeServer.ts — the OpenAI-compatible API bridge server.
- src/server/ws/liveServer.ts — the live-dashboard WebSocket daemon (20132).
- src/lib/services/embedWsProxy.ts — the embedded-service WS proxy (20131).

Each now calls attachRequestStreamGuards(req, res) inside its
http.createServer request listener, and installProcessCrashGuard() once
at start (idempotent process-level safety net).

The guard implementation is promoted to a single source of truth at
src/shared/utils/httpClientAbortGuard.mjs (importable from both the
Node-only dev server and the TypeScript servers; tsconfig allowJs:true).
scripts/dev/httpClientAbortGuard.mjs is now a pure re-export so the dev
server's existing import path is unchanged.

TDD: adds a test asserting the dev re-export is the SAME function objects
as the shared module (guards against future drift). Full guard suite stays
green (8/8).
